create the number list
Was thinking how to create the diceware number list so came up with this solution. I think this is the easiest way, at least for me ;)
It's just a list done with seq 11111 66666 and than removing the lines with 0s, 7s, 8s and 9s with grep -v -e 0 -e 7 -e 8 -e 9 tmp_numbers.asc 
Feel free to remove the extra delays and text displays decorations in the script, was just for extra joy.
Have fun!